Add factoring of inbound sidepanel
