+++ packages/invoices/src/InboundShipment/DetailView/SidePanel/SidePanel.tsx
@@ -1,0 +1,41 @@
+import React, { FC } from 'react';
+import {
+  CopyIcon,
+  DetailPanelAction,
+  DetailPanelPortal,
+  useNotification,
+  useTranslation,
+} from '@openmsupply-client/common';
+import { useInbound } from '../../api';
+import { AdditionalInfoSection } from './AdditionalInfoSection';
+import { PricingSection } from './PricingSection';
+import { RelatedDocumentsSection } from './RelatedDocumentsSection';
+
+export const SidePanel: FC = () => {
+  const { data } = useInbound();
+  const { success } = useNotification();
+  const t = useTranslation('common');
+
+  const copyToClipboard = () => {
+    navigator.clipboard.writeText(JSON.stringify(data, null, 4) ?? '');
+    success('Copied to clipboard successfully')();
+  };
+
+  return (
+    <DetailPanelPortal
+      Actions={
+        <>
+          <DetailPanelAction
+            icon={<CopyIcon />}
+            title={t('link.copy-to-clipboard')}
+            onClick={copyToClipboard}
+          />
+        </>
+      }
+    >
+      <AdditionalInfoSection />
+      <RelatedDocumentsSection />
+      <PricingSection />
+    </DetailPanelPortal>
+  );
+};
