+++ packages/invoices/src/InboundShipment/DetailView/SidePanel/index.ts
@@ -1,0 +1,1 @@
+export * from './SidePanel';
